Versão 2.2

+++ src/main/java/br/com/alelo/consumer/consumerpat/payload/ConsumerCardResponse.java
@@ -1,5 +1,7 @@
 package br.com.alelo.consumer.consumerpat.payload;
 
+import java.math.BigDecimal;
+
 import br.com.alelo.consumer.consumerpat.enumeration.CardType;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,6 +12,6 @@ public class ConsumerCardResponse{
 
 	private String cardNumber;
 	private CardType cardType;
-	private Double cardBalance;
+	private BigDecimal cardBalance;
 
 }
